requirements.txt
flask
python-dotenv
groq
gunicorn
flask-cors
